+++ native/cocos/renderer/gfx-base/GFXDef.h
@@ -32,11 +32,11 @@ namespace cc {
 namespace gfx {
 
 template <typename T, typename Enable = std::enable_if_t<std::is_class<T>::value>>
-struct Hasher final { size_t operator()(const T &info) const; };
+struct Hasher final { uint32_t operator()(const T &info) const; };
 
 // make this boost::hash compatible
 template <typename T, typename Enable = std::enable_if_t<std::is_class<T>::value>>
-size_t hash_value(const T &info) { return Hasher<T>()(info); } // NOLINT(readability-identifier-naming)
+uint32_t hash_value(const T &info) { return Hasher<T>()(info); } // NOLINT(readability-identifier-naming)
 
 #define DEFINE_CMP_OP(type)                                   \
     bool        operator==(const type &lhs, const type &rhs); \
