+++ native/cocos/renderer/gfx-base/GFXInputAssembler.h
@@ -43,7 +43,7 @@ public:
     inline const BufferList &   getVertexBuffers() const { return _vertexBuffers; }
     inline Buffer *             getIndexBuffer() const { return _indexBuffer; }
     inline Buffer *             getIndirectBuffer() const { return _indirectBuffer; }
-    inline size_t               getAttributesHash() const { return _attributesHash; }
+    inline uint32_t             getAttributesHash() const { return _attributesHash; }
 
     inline const DrawInfo &getDrawInfo() const { return _drawInfo; }
     inline void            setDrawInfo(const DrawInfo &info) { _drawInfo = info; }
@@ -68,10 +68,10 @@ protected:
     virtual void doInit(const InputAssemblerInfo &info) = 0;
     virtual void doDestroy()                            = 0;
 
-    size_t computeAttributesHash() const;
+    uint32_t computeAttributesHash() const;
 
     AttributeList _attributes;
-    size_t        _attributesHash = 0;
+    uint32_t      _attributesHash = 0;
 
     BufferList _vertexBuffers;
     Buffer *   _indexBuffer{nullptr};
