+++ native/cocos/renderer/gfx-metal/MTLUtils.mm
@@ -40,7 +40,7 @@ namespace cc {
 namespace gfx {
 namespace {
 
-ccstd::unordered_map<size_t, PipelineState*> pipelineMap;
+ccstd::unordered_map<uint32_t, PipelineState*> pipelineMap;
 ccstd::unordered_map<size_t, RenderPass*> renderPassMap;
 
 EShLanguage getShaderStage(ShaderStageFlagBit type) {
@@ -391,7 +391,7 @@ gfx::Shader *createShader(CCMTLDevice *device, CCMTLRenderPass* renderPass) {
 }
 
 CCMTLGPUPipelineState *getClearRenderPassPipelineState(CCMTLDevice *device, RenderPass * curPass) {
-    size_t rpHash = curPass->getHash();
+    uint32_t rpHash = curPass->getHash();
     const auto iter = pipelineMap.find(rpHash);
     if(iter != pipelineMap.end()) {
         auto *ccMtlPiplineState = static_cast<CCMTLPipelineState *>(iter->second);
